Update on Row (incomplete)
- Objects are now generated by Row
- Temp teachers are now added

+++ Underperforming-Students-vs-Teachers/timevariant.cpp
@@ -9,7 +9,7 @@ QLabel* TimeVariant::getWidget() {return widget;}
 
 void TimeVariant::setWidget(QLabel* widget) {this->widget = widget;}
 
-double TimeVariant::getDistanceFromLeft() const {
+int TimeVariant::getDistanceFromLeft() const {
     return this->widget->x();
 }
